+++ src/main/java/assignment/individualtrack/persistence/GameRepo.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface GameRepo extends JpaRepository<GameEntity, Integer> {
+public interface GameRepo extends JpaRepository<GameEntity, Long> {
     List<GameEntity> findTop20ByPlayerIdOrderByIdDesc(long playerId);
 }
