+++ src/main/java/assignment/individualtrack/persistence/ThemeRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface ThemeRepo extends JpaRepository<ThemeEntity, Integer> {
+public interface ThemeRepo extends JpaRepository<ThemeEntity, Long> {
     boolean existsByName(String name);
     Optional<ThemeEntity> findById(Long id);
     void deleteById(Long id);
